chore: space for partial impls (ugly copy paste again)
